+++ app/models/compromise_key.rb
@@ -1,0 +1,3 @@
+class CompromiseKey < RejectKey
+    
+end