+++ db/migrate/20200731200909_modify_weak_keys.rb
@@ -1,0 +1,8 @@
+class ModifyWeakKeys < ActiveRecord::Migration
+  def change
+    rename_table :weak_keys, :reject_keys
+    rename_column :reject_keys, :sha1_hash, :fingerprint 
+    add_column :reject_keys, :source, :string
+    add_column :reject_keys, :type, :string
+  end
+end